config_flow: Print exception info to log
We currently consume every `OmnikInverterError` and only raise a
`cannot_connect` error to the UI, with no context whatsoever.  It could
be a connection error, but also a message that fails to parse/verify.
Print the stacktrace to the log so that we can better aid our users when
they encounter issues; this also makes it easier to understand and solve
compatibility problems with other brands that use a very similar yet not
identical message format.